Split main mint projects class into four distint classes
