+++ src/main/java/com/everneth/emi/EMI.java
@@ -5,7 +5,10 @@ import co.aikar.idb.*;
 import com.everneth.emi.commands.*;
 import com.everneth.emi.commands.bot.*;
 import com.everneth.emi.commands.bot.par.WhitelistAppCommand;
-import com.everneth.emi.commands.mint.DevopCommand;
+import com.everneth.emi.commands.mint.projects.MintMaterialCommands;
+import com.everneth.emi.commands.mint.projects.MintProjectCommands;
+import com.everneth.emi.commands.mint.projects.MintTaskCommands;
+import com.everneth.emi.commands.mint.projects.MintValidationCommands;
 import com.everneth.emi.commands.par.CharterCommand;
 import com.everneth.emi.events.JoinEvent;
 import com.everneth.emi.events.LeaveEvent;
@@ -107,7 +110,10 @@ public class EMI extends JavaPlugin {
         commandManager.registerCommand(new SupportCommand());
         commandManager.registerCommand(new DiscordsyncCommand());
         commandManager.registerCommand(new CharterCommand());
-        commandManager.registerCommand(new DevopCommand());
+        commandManager.registerCommand(new MintProjectCommands());
+        commandManager.registerCommand(new MintMaterialCommands());
+        commandManager.registerCommand(new MintTaskCommands());
+        commandManager.registerCommand(new MintValidationCommands());
         commandManager.registerCommand(new MotdCommand());
     }
